requirements.txt
flask
line-bot-sdk
ibm-watson>=4.3.0
